Cambios en telemetría
Facade debería retornar solo un dataframe con los valores (ver diagrama de clases)
